Change: all links of all occurences of .adminmenu got a span.icon
The use of an element for the icons makes it easier to display them in different sizes.
The sizes of the icons *inside* the span.icon are defined *once*,
The different sizes of the icons are defined with only different values of width and height of .icon.
Additionally all backup options got their own icons. This includes the new icon clock.svg.