Fix: prevent passwords from being spell checked in case of being shown
When a user displays the password in its input, the forum software does it with changing the input type from `password` to `text`. In this moment it is possible, that the browser sends the content (the password) to an external service (in example Chrome sends the content to a Google service) for spell checking reasons.